refactor(sdk-core): tss refactor eddsa into a folder so ecdsa can be added
Ticket: BG-49287
