+++ modules/sdk-core/src/bitgo/tss/eddsa/index.ts
@@ -1,3 +1,9 @@
+import * as EDDSAMethods from './eddsa';
+export * as EDDSAMethodTypes from './types';
+
+export default EDDSAMethods;
+
+// exporting this types for backward compatibility.
 export {
   createCombinedKey,
   createUserSignShare,
@@ -9,3 +15,5 @@ export {
   sendSignatureShare,
   encryptYShare,
 } from './eddsa';
+
+export { EncryptedYShare, DecryptableYShare, CombinedKey, SigningMaterial } from './types';
