+++ modules/sdk-core/src/bitgo/utils/index.ts
@@ -3,12 +3,11 @@ import * as openpgpUtils from './opengpgUtils';
 export * from './abstractUtxoCoinUtil';
 export * from './blsUtils';
 export * from './iBlsUtils';
-export * from './iTssUtils';
 export * from './mpcUtils';
 export * from './opengpgUtils';
 export * from './promise-utils';
 export * from './triple';
-export * from './tssUtils';
+export * from './tss';
 export * from './util';
 
 export { openpgpUtils };
