+++ modules/sdk-core/src/bitgo/utils/tss/index.ts
@@ -1,0 +1,12 @@
+import TssUtils, { TssUtilsTypes } from './eddsa';
+export { TssUtils, TssUtilsTypes };
+
+// exporting this types for backward compatibility.
+export {
+  ITssUtils,
+  PrebuildTransactionWithIntentOptions,
+  SignatureShareRecord,
+  SignatureShareType,
+  TxRequest,
+  UnsignedTransaction,
+} from './eddsa';
